refactor: remove form into separate folder
